fix: landing page error By moving the styles file to another folder
